handle missing python2 gracefully
If python2 is missing, git-push-to-hg will fail after already modified the
mercurial repo state. So abort gracefully before starting on such case.

Scripts which are python(2) themselves will fail before starting. Currently
git-push-to-hg is the only bash script which invokes a python(2) script.

Also add a note at the README about handling it with Windows mozilla-build.